Trying user-agent header to set app name, OS name/version
Setting header info isn't too hard - create a session and use session.post
instead of requests.post.

Not sure if this will work though because we don't know how GA processes
the user-agent string. If it has hard templates based on existing browsers
then we'll need to pretend to be a browser. If not then maybe we can use
our own app name and see if it will scrape the info from the string.